+++ client/web/src/enterprise/insights/core/backend/gql-backend/methods/get-insight-preview.ts
@@ -76,7 +76,7 @@ export const getInsightsPreview = (
             const seriesMetadata = indexedSeries.map((generatedSeries, index) => ({
                 id: generatedSeries.seriesId,
                 name: generatedSeries.label,
-                query: inputMetadata[generatedSeries.label]?.query || '',
+                query: inputMetadata[`${generatedSeries.label}-${index}`]?.query || '',
                 stroke: getColorForSeries(generatedSeries.label, index),
             }))
 
@@ -91,9 +91,10 @@ export const getInsightsPreview = (
                         value: point.value,
                         dateTime: new Date(point.dateTime),
                         link: generateLinkURL({
-                            previousPoint: line.points[index - 1],
-                            series: seriesDefinitionMap[line.seriesId],
                             point,
+                            previousPoint: line.points[index - 1],
+                            query: seriesDefinitionMap[line.seriesId].query,
+                            repositories: input.repositories,
                         }),
                     })),
                     name: line.label,